+++ src/lib/firebase.ts
@@ -29,6 +29,7 @@ const firebaseConfig = {
 }
 
 const app = initializeApp(firebaseConfig)
+export const firebaseApp = app
 export const auth = getAuth(app)
 export const db = getFirestore(app)
 const googleProvider = new GoogleAuthProvider()